fix: clean up and document Transaction class implementation
- Add missing final modifiers to Type enum fields
- Remove @Override annotations from non-inherited methods
- Fix toString() formatting and padding inconsistencies
- Add comprehensive Doxygen documentation to all classes and methods
- Remove duplicate account implementation file
